refactor(hw/bios): moved to hw/bios/*
